+++ glfw/glfw.py
@@ -323,7 +323,7 @@ def generate_wrappers(glfw_header: str) -> None:
     void glfwWaylandRunWithActivationToken(GLFWwindow *handle, GLFWactivationcallback cb, void *cb_data)
     bool glfwWaylandSetTitlebarColor(GLFWwindow *handle, uint32_t color, bool use_system_color)
     void glfwWaylandRedrawCSDWindowTitle(GLFWwindow *handle)
-    void glfwWaylandSetupLayerShellForNextWindow(GLFWLayerShellConfig c)
+    void glfwWaylandSetupLayerShellForNextWindow(const GLFWLayerShellConfig *c)
     pid_t glfwWaylandCompositorPID(void)
     unsigned long long glfwDBusUserNotify(const char *app_name, const char* icon, const char *summary, const char *body, \
 const char *action_text, int32_t timeout, int urgency, GLFWDBusnotificationcreatedfun callback, void *data)